Fix navigation bug
When user selects Answered questions, they should be navigated to Results page and should be able to view the results of the page.
Fix: Passed a prop from QuestionList to QuestionList item which is a boolean value. If this value is ytrue then the question is answered otherwise it is unanswered.
If the question is answered then navigate to the poll results page directly otherwise first navigate to the Questions page and submit the answer.